[PRMT-3415] Renamed ehr-deletion-job-common.js.
