+++ src/services/jobs/ehr-deletion-job/delete-from-repo-db.js
@@ -1,9 +1,9 @@
-import { S3Service } from '../../storage';
 import { hardDeleteAllMessagesByConversationId } from '../../database/message-repository';
 import { hardDeleteHealthRecordByConversationId } from '../../database/health-record-repository';
 import { setCurrentSpanAttributes } from '../../../config/tracing';
 import { logError, logInfo } from '../../../middleware/logging';
-import { loggerPrefix } from './ehr-deletion-job-utilities';
+import { loggerPrefix } from './ehr-deletion-job-common';
+import { S3Service } from '../../storage';
 
 export const permanentlyDeleteEhrFromRepoAndDb = async (healthRecord) => {
   const { conversationId } = healthRecord;
