+++ src/services/jobs/ehr-deletion-job/ehr-deletion-job-common.js
@@ -1,0 +1,1 @@
+export const loggerPrefix = '[SCHEDULED JOB] [EHR DELETION] -';
